Refactor OSITrace further and add optional caching
Also renames the module file to fit PEP-8 conventions.
